requirements.txt
pytest~=6.2.5
pytest-selenium
termcolor~=2.1.1
allure-python-commons
selenium~=4.7.2
beautifulsoup4~=4.11.1
2captcha-python~=1.1.3